refactor(db): extract queries into domain repository modules (#1405)
Replaces the monolithic backend/src/db.rs (1,964 lines) with five
domain-focused modules, zero API surface changes, and new connection
pool metrics.

New layout:
  backend/src/db/
  ├── mod.rs          pool creation + retry + re-exports (305 lines)
  ├── pools.rs        pool CRUD, templates, creator incentives (580 lines)
  ├── predictions.rs  prediction queries, leaderboard, stats (420 lines)
  ├── referrals.rs    referral inserts and earnings (130 lines)
  ├── metrics.rs      connection pool Prometheus gauges (190 lines)
  └── README.md       architecture documentation

Backwards compatibility:
  All 32 public functions and types are re-exported from db/mod.rs.
  Existing crate::db::* call sites require zero changes.

New connection pool metrics (closes #1405):
  db_pool_size, db_pool_idle, db_pool_active, db_pool_utilization_ratio
  Registered via db::metrics::register_pool_gauges() at startup.
  Updated via PoolMetricGauges::record(&pool) in a background task.

Duplication eliminated:
  - calculate_odds extracted from 2 inline sites into pools.rs
  - Rank offset pattern unified across 3 leaderboard queries
  - Bulk-insert placeholder builder extracted to referrals::insert_referrals_chunk
  - Win-rate divide-by-zero guard consolidated in predictions.rs

Net diff: -339 lines (1,625 added across 5 new files, 1,964 removed)
Tests:    all existing tests preserved + co-located with queries

Closes #1405
